fix nonsensical `operate` test set
Currently the test set starts with the line
`@testset "$(typeof(x))" for x in [2, big(3)]`, which points to the
test not working as intended, because the array constructor
(`[2, big(3)]`) promotes all elements to a common type, so naming
the test set `"$(typeof(x))"` doesn't make sense.

Furthermore, after the above is fixed, the test fails because the `!==`
check is performed independently of whether the type is mutable.

TODO: including `Rational{BigInt}` among the tested types reveals what
seems to be a bug: #240.

TODO: including `abs` among the tested operations reveals that it lacks
an `operate` method: #241.